requirements.txt
colorama
keras-cv
fastapi
matplotlib
numpy
pandas
opencv-python
Pillow
pillow-heif
python-multipart
requests
starlette
streamlit
tensorflow
ultralytics
uvicorn
